hide more implementation from public interface of morphology
