+++ arbor/include/arbor/morph/morphology.hpp
@@ -10,23 +10,6 @@
 
 namespace arb {
 
-// An unbranched cable segment that has root, terminal or fork point at each end.
-struct mbranch {
-    std::vector<msize_t> index;  // sample index
-    msize_t parent_id = mnpos;   // branch index
-
-    mbranch() = default;
-    mbranch(std::vector<msize_t> idx, msize_t parent):
-        index(std::move(idx)), parent_id(parent) {}
-
-    bool is_sphere()  const { return size()==1u; }
-    msize_t size()    const { return index.size(); }
-    bool has_parent() const { return parent_id!=mnpos;}
-
-    friend bool operator==(const mbranch& l, const mbranch& r);
-    friend std::ostream& operator<<(std::ostream& o, const mbranch& b);
-};
-
 class morphology_impl;
 
 using mindex_range = std::pair<const msize_t*, const msize_t*>;
