+++ arbor/morph/morphology.cpp
@@ -11,12 +11,28 @@
 
 #include "algorithms.hpp"
 #include "io/sepval.hpp"
+#include "morph/morphology_impl.hpp"
 #include "util/span.hpp"
 #include "util/strprintf.hpp"
 
 namespace arb {
 
 namespace impl{
+
+//
+//  mbranch implementation
+//
+
+bool operator==(const mbranch& l, const mbranch& r) {
+    return l.parent_id==r.parent_id && l.index==r.index;
+}
+
+std::ostream& operator<<(std::ostream& o, const mbranch& b) {
+    o <<"mbranch([" << io::csv(b.index) << "], ";
+    if (b.parent_id==mnpos) o << "none)";
+    else  o << b.parent_id << ")";
+    return o;
+}
 
 std::vector<mbranch> branches_from_parent_index(const std::vector<msize_t>& parents,
                                                 const std::vector<point_prop>& props,
@@ -94,21 +110,6 @@ bool root_sample_has_same_tag_as_child(const sample_tree& st) {
 } // namespace impl
 
 //
-//  mbranch implementation
-//
-
-bool operator==(const mbranch& l, const mbranch& r) {
-    return l.parent_id==r.parent_id && l.index==r.index;
-}
-
-std::ostream& operator<<(std::ostream& o, const mbranch& b) {
-    o <<"mbranch([" << io::csv(b.index) << "], ";
-    if (b.parent_id==mnpos) o << "none)";
-    else  o << b.parent_id << ")";
-    return o;
-}
-
-//
 //  morphology_impl definition and implementation
 //
 
@@ -120,7 +121,7 @@ struct morphology_impl {
     bool spherical_root_ = false;
 
     // Branch state.
-    std::vector<mbranch> branches_;
+    std::vector<impl::mbranch> branches_;
     std::vector<msize_t> branch_parents_;
     std::vector<std::vector<msize_t>> branch_children_;
